moved RotarySelector test to own file
