+++ firmware/common2015/testing/ExampleTest.cpp
@@ -1,0 +1,4 @@
+#include <gtest/gtest.h>
+
+// Example unit test
+TEST(ExampleTest, example) { EXPECT_EQ(true, !false); }
